+++ library/src/index.ts
@@ -2,5 +2,6 @@ export * from './error/index.ts';
 export * from './methods/index.ts';
 export * from './schemas/index.ts';
 export * from './transformations/index.ts';
+export * from './utils/index.ts';
 export * from './validations/index.ts';
 export * from './types.ts';