+++ library/src/methods/keyof/keyof.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
+import { comparable } from '../../comparable.ts';
 import { enumType, number, object, string } from '../../schemas/index.ts';
-import { comparable } from '../../utils/index.ts';
 import { keyof } from './keyof.ts';
 
 describe('keyof', () => {