+++ library/src/methods/pick/pickAsync.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { comparable } from '../../utils/index.ts';
+import { comparable } from '../../comparable.ts';
 import { object, objectAsync, string } from '../../schemas/index.ts';
 import { parseAsync } from '../parse/index.ts';
 import { pickAsync } from './pickAsync.ts';