+++ library/src/utils/comparable/index.ts
@@ -1,1 +1,0 @@
-export * from './comparable.ts';