+++ library/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './comparable/index.ts';
 export * from './executePipe/index.ts';
 export * from './getCurrentPath/index.ts';
 export * from './getErrorAndPipe/index.ts';